ingestion/requirements.txt
boto3>=1.28.0
awswrangler>=3.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
pandas>=2.0.0
pyarrow>=12.0.0